fmu-runner/requirements.txt
quixstreams==3.23.2
fmpy>=0.3.0
numpy
pandas
python-dotenv
requests

# Testing
pytest>=7.0.0
pytest-cov
